+++ app/src/main/java/com/example/compose_todo/MainActivity.kt
@@ -13,6 +13,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
+import com.example.compose_todo.UIcomponents.MainScreen
+import com.example.compose_todo.UIcomponents.NoteUI
 import com.example.compose_todo.ui.theme.ComposetodoTheme
